feat(deal-ticket): hide price on market orders
