+++ apps/trading-e2e/src/integration/trading-deal-ticket-order.cy.ts
@@ -64,7 +64,7 @@ describe('deal ticker order validation', { tags: '@smoke' }, () => {
       cy.getByTestId(orderSizeField).clear().type('1');
       cy.getByTestId(orderPriceField).clear().type('1.123456');
       cy.getByTestId(placeOrderBtn).click();
-      cy.getByTestId('deal-ticket-error-message-price-limit').should(
+      cy.getByTestId('deal-ticket-error-message-price').should(
         'have.text',
         'Price accepts up to 5 decimal places'
       );
@@ -87,7 +87,7 @@ describe('deal ticker order validation', { tags: '@smoke' }, () => {
       cy.getByTestId(orderSizeField).clear().type('1.234');
       // 7002-SORD-060
       cy.getByTestId(placeOrderBtn).should('be.enabled');
-      cy.getByTestId('deal-ticket-error-message-size-market').should(
+      cy.getByTestId('deal-ticket-error-message-size').should(
         'have.text',
         'Size must be whole numbers for this market'
       );
@@ -96,7 +96,7 @@ describe('deal ticker order validation', { tags: '@smoke' }, () => {
     it('must warn if order size is set to 0', function () {
       cy.getByTestId(orderSizeField).clear().type('0');
       cy.getByTestId(placeOrderBtn).should('be.enabled');
-      cy.getByTestId('deal-ticket-error-message-size-market').should(
+      cy.getByTestId('deal-ticket-error-message-size').should(
         'have.text',
         'Size cannot be lower than 1'
       );
